fix: search formatting
