+++ src/components/BlogRoll/constants.ts
@@ -1,0 +1,2 @@
+// Don't show search results below this char count
+export const SEARCH_THRESHOLD = 2;
