+++ src/components/BlogRoll/types.ts
@@ -1,0 +1,17 @@
+export type Post = {
+  excerpt?: string;
+  rawMarkdownBody: string;
+  frontmatter: {
+    title: string;
+    description: string;
+    displaytitle: string;
+    featuredimage: string;
+  };
+  fields: {
+    slug: string;
+  };
+};
+
+export type PostEdge = {
+  node: Post;
+};
